Update README.md
This project involves the development of a complete website, which I built entirely on my own, with both front-end and back-end functionalities. The website is designed to be a fully functional and user-friendly web application, where I have implemented a variety of features, such as:

Front-end Development: I used HTML, CSS, and JavaScript to create a responsive and interactive user interface. The front-end is designed with attention to both aesthetics and usability, ensuring that the user experience is seamless across devices.
Back-end Development: The back-end is built using PHP. This side handles the logic, database management, and communication between the server and the client-side. I implemented.
Database Integration: The website uses MySQL to store and manage user data, enabling dynamic content and smooth interaction with the website.
The website is accessible globally.

I made sure to follow best practices for both front-end and back-end development to ensure scalability, security, and performance. The project was a great opportunity for me to enhance my skills in full-stack development and demonstrate my ability to independently build and deploy a complete web application.